Remove all top-level code.
This lets the module's source code be used as-os a in a regular app being developed with Xcode. By default, Xcode 7.3 uses the magic @NSApplicationMain attribute. If there is any top-level code anywhere in any of the other source code files, this error occurs:

'NSApplicationMain' attribute cannot be used in a module that contains top-level code

I want to play with the experimental Swift Package Manager, but I also want to use the source code in this package with regular apps based on the pre-package-manager Xcode.